Adding a data repository
I have added a way to encapsulate how data is store for the mocked client.  I have decided to go with litedb to give the project simple document storage.
This will be interface out, with the implementation being passed into the main entry point so that new providers can be written if needed.